--- core/app-framework/base/native/base_lib_export.h
@@ -1,13 +1,0 @@
-/*
- * Copyright (C) 2019 Intel Corporation.  All rights reserved.
- * SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
- */
-
-#ifndef _BASE_LIB_EXPORT_H_
-#define _BASE_LIB_EXPORT_H_
-
-#include "bi-inc/attr_container.h"
-#include "native_interface.h"
-
-#endif /* end of _BASE_LIB_EXPORT_H_ */
-
