--- core/app-framework/sensor/native/sensor_api.h
@@ -1,32 +1,0 @@
-/*
- * Copyright (C) 2019 Intel Corporation.  All rights reserved.
- * SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
- */
-
-#ifndef _SENSOR_API_H_
-#define _SENSOR_API_H_
-
-#include "bh_platform.h"
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-uint32
-wasm_sensor_open(int32 name_offset, int instance);
-
-bool
-wasm_sensor_config(uint32 sensor, int interval, int bit_cfg, int delay);
-
-bool
-wasm_sensor_config_with_attr_container(uint32 sensor, int32 buffer_offset, int len);
-
-bool
-wasm_sensor_close(uint32 sensor);
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif /* end of _SENSOR_API_H_ */
-
