+++ core/app-framework/sensor/native/sensor_native_api.h
@@ -1,0 +1,37 @@
+/*
+ * Copyright (C) 2019 Intel Corporation.  All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+ */
+
+#ifndef _SENSOR_NATIVE_API_H_
+#define _SENSOR_NATIVE_API_H_
+
+#include "bh_platform.h"
+#include "wasm_export.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+bool
+wasm_sensor_config(wasm_exec_env_t exec_env,
+                   uint32 sensor, int interval,
+                   int bit_cfg, int delay);
+uint32
+wasm_sensor_open(wasm_exec_env_t exec_env,
+                 int32 name_offset, int instance);
+
+bool
+wasm_sensor_config_with_attr_container(wasm_exec_env_t exec_env,
+                                       uint32 sensor, int32 buffer_offset,
+                                       int len);
+
+bool
+wasm_sensor_close(wasm_exec_env_t exec_env, uint32 sensor);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* end of _SENSOR_NATIVE_API_H_ */
+
