+++ samples/gui/wasm-runtime-wgl/src/ext_lib_export.c
@@ -1,7 +1,7 @@
 #include "lib_export.h"
-#include "sensor_api.h"
-#include "connection_api.h"
-#include "gui_api.h"
+#include "sensor_native_api.h"
+#include "connection_native_api.h"
+#include "gui_native_api.h"
 
 static NativeSymbol extended_native_symbol_defs[] = {
 #include "runtime_sensor.inl"
