+++ samples/simple/src/ext_lib_export.c
@@ -1,6 +1,8 @@
 #include "lib_export.h"
-#include "sensor_api.h"
-#include "connection_api.h"
+#include "sensor_native_api.h"
+#include "timer_native_api.h"
+#include "req_resp_native_api.h"
+#include "connection_native_api.h"
 
 static NativeSymbol extended_native_symbol_defs[] = {
 #include "runtime_sensor.inl"
